Move Statuses to the new Statuses subnamespace
