+++ site/app/Training/Applications/TrainingApplication.php
@@ -8,7 +8,7 @@ use MichalSpacekCz\Training\Files\TrainingFiles;
 use MichalSpacekCz\Training\Files\TrainingFilesCollection;
 use MichalSpacekCz\Training\Mails\MailMessageAdmin;
 use MichalSpacekCz\Training\Mails\TrainingMailMessageFactory;
-use MichalSpacekCz\Training\Statuses;
+use MichalSpacekCz\Training\Statuses\Statuses;
 use Nette\Utils\Html;
 
 class TrainingApplication
