+++ site/config/services.neon
@@ -137,7 +137,7 @@ services:
 	- MichalSpacekCz\Training\Resolver\Vrana
 	- MichalSpacekCz\Training\Reviews\TrainingReviewInputsFactory
 	- MichalSpacekCz\Training\Reviews\TrainingReviews
-	- MichalSpacekCz\Training\Statuses
+	- MichalSpacekCz\Training\Statuses\Statuses
 	- MichalSpacekCz\Training\Statuses\TrainingStatusHistory
 	- MichalSpacekCz\Training\TrainingLocales
 	- MichalSpacekCz\Training\Trainings\TrainingFactory
