+++ site/tests/Training/Mails/TrainingMailMessageFactoryTest.phpt
@@ -9,7 +9,7 @@ use MichalSpacekCz\Test\Database\Database;
 use MichalSpacekCz\Test\TestCaseRunner;
 use MichalSpacekCz\Training\Applications\TrainingApplication;
 use MichalSpacekCz\Training\Files\TrainingFiles;
-use MichalSpacekCz\Training\Statuses;
+use MichalSpacekCz\Training\Statuses\Statuses;
 use Nette\Utils\Html;
 use Tester\Assert;
 use Tester\TestCase;
